#1 Refactor VPC service and relevant class. The AcsClientStub now is registered as a service.
